+++ Form1.cs
@@ -8,11 +8,11 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace GroupAssignment
+namespace Prototype
 {
-    public partial class FormRegisterUser : Form
+    public partial class formLogin : Form
     {
-        public FormRegisterUser()
+        public formLogin()
         {
             InitializeComponent();
         }
@@ -27,41 +27,15 @@ namespace GroupAssignment
 
         }
 
-        private void label3_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void label4_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void label5_Click(object sender, EventArgs e)
-        {
-
-        }
-
         private void label2_Click(object sender, EventArgs e)
         {
 
         }
 
-        private void textBox1_TextChanged(object sender, EventArgs e)
+        private void buttonLogin_Click(object sender, EventArgs e)
         {
+          
 
-        }
-
-        private void textBox4_TextChanged(object sender, EventArgs e)
-        {
-
-        }
-
-        private void buttonRegister_Click(object sender, EventArgs e)
-        {
-            MessageBox.Show("Hello" + ' ' + textBoxFirstName.Text + ' ' + textBoxSurname.Text + ' ' +
-           " Welcome to Limerick Gamers Game Library ", "Thank You For Registering",  MessageBoxButtons.OKCancel,
-            MessageBoxIcon.Information); 
         }
     }
 }
