+++ Prototype1/Form1.cs
@@ -1,0 +1,67 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace GroupAssignment
+{
+    public partial class FormRegisterUser : Form
+    {
+        public FormRegisterUser()
+        {
+            InitializeComponent();
+        }
+
+        private void Form1_Load(object sender, EventArgs e)
+        {
+
+        }
+
+        private void label1_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        private void label3_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        private void label4_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        private void label5_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        private void label2_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        private void textBox1_TextChanged(object sender, EventArgs e)
+        {
+
+        }
+
+        private void textBox4_TextChanged(object sender, EventArgs e)
+        {
+
+        }
+
+        private void buttonRegister_Click(object sender, EventArgs e)
+        {
+            MessageBox.Show("Hello" + ' ' + textBoxFirstName.Text + ' ' + textBoxSurname.Text + ' ' +
+           " Welcome to Limerick Gamers Game Library ", "Thank You For Registering",  MessageBoxButtons.OKCancel,
+            MessageBoxIcon.Information); 
+        }
+    }
+}
