+++ Prototype1/Program.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace Prototype
+namespace GroupAssignment
 {
     static class Program
     {
@@ -16,7 +16,7 @@ namespace Prototype
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new formLogin());
+            Application.Run(new FormRegisterUser());
         }
     }
 }
